requirements.txt
mutagen
pycryptodome
wasmer
python-magic-bin
wasmer_compiler_cranelift
pyinstaller